feat: first steps into main page
