+++ decide/authentication/templates/login.html
@@ -92,7 +92,7 @@
         </form>
         <div class="google-login">
             <button class="google-button">
-                <a href="{% provider_login_url 'google'%}?next=http://127.0.0.1:8000/authentication/menu">
+                <a href="{% provider_login_url 'google'%}?next=http://127.0.0.1:8000/base/">
                     Login with Google
                 </a>
             </button>
