+++ decide/authentication/urls.py
@@ -1,7 +1,7 @@
 from django.urls import include, path
 from rest_framework.authtoken.views import obtain_auth_token
 
-from .views import MenuView, SigninView, GetUserView, LogoutView, RegisterView
+from .views import SigninView, GetUserView, LogoutView, RegisterView
 
 
 urlpatterns = [
@@ -10,6 +10,5 @@ urlpatterns = [
     path('getuser/', GetUserView.as_view()),
     path('register/', RegisterView.as_view(),name='register'),
     path('signin/', SigninView.as_view(), name='signin'),
-    path('menu', MenuView.as_view(), name='menuSignin'),
     path('', include('allauth.urls')),
 ]
