+++ decide/authentication/views.py
@@ -7,15 +7,15 @@ from rest_framework.status import (
 from rest_framework.views import APIView
 from rest_framework.authtoken.models import Token
 from django.contrib.auth.models import User
-from django.shortcuts import render
 
 from django.db import IntegrityError
-from django.shortcuts import get_object_or_404, render,redirect
+from django.shortcuts import get_object_or_404, render, redirect
 from django.contrib.auth.forms import UserCreationForm
 from django.core.exceptions import ObjectDoesNotExist
 from django.contrib.auth.views import LoginView
 from django.views.generic import TemplateView
 from django.contrib.auth import authenticate, login
+from django.urls import reverse
 
 from .forms import LoginForm
 
@@ -32,7 +32,9 @@ class SigninView(TemplateView):
             user = authenticate(request, username=username, password=password)
             if user is not None:
                 login(request, user)
-                return redirect("menuSignin")
+                print("He llegado a aquí")
+                menu_signin_url = reverse('menuSignin')
+                return redirect(menu_signin_url)
             else:
                 error_message="Usuario y/o contraseña incorrecto/a"
         else:
@@ -43,14 +45,6 @@ class SigninView(TemplateView):
     def get(self, request):
         form_class = LoginForm(None)
         return render(request, 'login.html', {'form': form_class, 'msg': None})
-    
-class MenuView(TemplateView):
-    def post(self, request):
-        return render(request, 'menu.html')
-    
-    def get_template_names(self):
-        return ['menu.html']
-
 
 class GetUserView(APIView):
     def post(self, request):
@@ -72,8 +66,6 @@ class LogoutView(APIView):
 
 
 class RegisterView(APIView):
-    
-  
 
     def get(self, request):
