+++ decide/base/static/base/style.css
@@ -1,0 +1,21 @@
+body {
+    margin: 0;
+    padding: 0;
+    padding-top: 80px;
+}
+
+header {
+    position: fixed;
+    width: 100%;
+    top: 0;
+}
+
+footer {
+    position: fixed;
+    width: 100%;
+    bottom: 0;
+}
+
+.bg-azul-personalizado {
+    background-color: #001e36;
+}
