+++ decide/base/templates/base.html
@@ -1,3 +1,5 @@
+
+
 <!DOCTYPE html>
 <html>
 <head>
@@ -20,6 +22,7 @@
     </style>
     {% block extrahead %}{% endblock %}
 </head>
+{% include "header.html" %}
 <body>
     {% block content %}
     {% endblock %}
