+++ decide/base/urls.py
@@ -1,1 +1,6 @@
-urlpatterns = []
+from django.urls import path
+from .views import MenuView
+
+urlpatterns = [ 
+    path('', MenuView.as_view(), name='menuSignin'),
+]
