+++ decide/base/views.py
@@ -1,3 +1,9 @@
 from django.shortcuts import render
+from django.views.generic import TemplateView
 
-# Create your views here.
+class MenuView(TemplateView):
+    def post(self, request):
+        return render(request, 'menu.html')
+    
+    def get_template_names(self):
+        return ['menu.html']
